requirements.txt
pandas
numpy
sqlalchemy
pymysql
mysql-connector-python
openmeteo_requests
requests_cache
retry_requests
great_expectations